requirements.txt
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.6.0
billiard==3.3.0.23
celery==3.1.25
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
kombu==3.0.37
MarkupSafe==1.0
nose==1.3.7
pytz==2018.5
redis==2.10.6
requests==2.19.1
unbabel-py==0.49
urllib3==1.23
Werkzeug==0.14.1
